move deprecation test to correct 1.6 test file
