+++ tests/deprecated_api/test_remove_1-4.py
@@ -15,9 +15,7 @@
 
 import pytest
 
-from pytorch_lightning import Trainer
 from tests.deprecated_api import _soft_unimport_module
-from tests.helpers import BoringModel
 
 
 def test_v1_4_0_deprecated_imports():
@@ -25,15 +23,3 @@ def test_v1_4_0_deprecated_imports():
     with pytest.deprecated_call(match='will be removed in v1.4'):
         from pytorch_lightning.utilities.argparse_utils import _gpus_arg_default  # noqa: F811 F401
 
-
-def test_v1_4_0_deprecated_hpc_load(tmpdir):
-    model = BoringModel()
-    trainer = Trainer(
-        default_root_dir=tmpdir,
-        max_steps=1,
-    )
-    trainer.fit(model)
-    trainer.checkpoint_connector.hpc_save(tmpdir, trainer.logger)
-    checkpoint_path = trainer.checkpoint_connector.get_max_ckpt_path_from_folder(str(tmpdir))
-    with pytest.deprecated_call(match=r"`CheckpointConnector.hpc_load\(\)` was deprecated in v1.4"):
-        trainer.checkpoint_connector.hpc_load(checkpoint_path)
