+++ tests/deprecated_api/test_remove_1-6.py
@@ -309,3 +309,16 @@ def test_v1_6_0_deprecated_disable_validation():
 def test_v1_6_0_every_n_val_epochs():
     with pytest.deprecated_call(match="use `every_n_epochs` instead"):
         _ = ModelCheckpoint(every_n_val_epochs=1)
+
+
+def test_v1_4_0_deprecated_hpc_load(tmpdir):
+    model = BoringModel()
+    trainer = Trainer(
+        default_root_dir=tmpdir,
+        max_steps=1,
+    )
+    trainer.fit(model)
+    trainer.checkpoint_connector.hpc_save(tmpdir, trainer.logger)
+    checkpoint_path = trainer.checkpoint_connector.get_max_ckpt_path_from_folder(str(tmpdir))
+    with pytest.deprecated_call(match=r"`CheckpointConnector.hpc_load\(\)` was deprecated in v1.4"):
+        trainer.checkpoint_connector.hpc_load(checkpoint_path)
